widget_project/requirements.txt
asgiref==3.3.4
autopep8==1.5.7
bandit==1.7.0
Django==3.2.3
djangorestframework==3.12.4
flake8==3.9.2
gitdb==4.0.7
GitPython==3.1.17
mccabe==0.6.1
pbr==5.6.0
pycodestyle==2.7.0
pyflakes==2.3.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
smmap==4.0.0
sqlparse==0.4.1
stevedore==3.3.0
time-machine==2.1.0
toml==0.10.2
